[IDP-2296] remove interface for TransactionConsumer
